// server_2x/src/octoprint/dto/websocket-output/timelapse-message.dto.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface TimelapseMessageDto {
  timelapse?: null;
  [k: string]: any;
}