// server_2x/src/octoprint/dto/websocket-output/current-message.dto.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface CurrentMessageDto {
  current?: {
    state?: {
      text?: string;
      flags?: {
        operational?: boolean;
        printing?: boolean;
        cancelling?: boolean;
        pausing?: boolean;
        resuming?: boolean;
        finishing?: boolean;
        closedOrError?: boolean;
        error?: boolean;
        paused?: boolean;
        ready?: boolean;
        sdReady?: boolean;
        [k: string]: any;
      };
      [k: string]: any;
    };
    job?: {
      file?: {
        name?: null;
        path?: null;
        display?: null;
        origin?: null;
        size?: null;
        date?: null;
        [k: string]: any;
      };
      estimatedPrintTime?: null;
      averagePrintTime?: null;
      lastPrintTime?: null;
      filament?: null;
      user?: null;
      [k: string]: any;
    };
    currentZ?: null;
    progress?: {
      completion?: null;
      filepos?: null;
      printTime?: null;
      printTimeLeft?: null;
      printTimeLeftOrigin?: null;
      [k: string]: any;
    };
    offsets?: {
      [k: string]: any;
    };
    resends?: {
      count?: number;
      transmitted?: number;
      ratio?: number;
      [k: string]: any;
    };
    serverTime?: number;
    temps?: {
      time?: number;
      tool0?: {
        actual?: number;
        target?: number;
        [k: string]: any;
      };
      bed?: {
        actual?: number;
        target?: number;
        [k: string]: any;
      };
      chamber?: {
        actual?: null;
        target?: null;
        [k: string]: any;
      };
      [k: string]: any;
    }[];
    logs?: string[];
    messages?: string[];
    busyFiles?: any[];
    [k: string]: any;
  };
  [k: string]: any;
}